pyproject.toml
[project]
name = "koclaw"
version = "0.1.0"
description = "The open-source AI agent built for Korea"
requires-python = ">=3.12"
license = { text = "MIT" }
dependencies = [
    "anthropic>=0.40.0",
    "openai>=1.50.0",
    "google-genai>=1.0.0",
    "aiosqlite>=0.20.0",
    "httpx>=0.27.0",
    "docker>=7.0.0",
    "slack-bolt>=1.21.0",
    "aiohttp>=3.9.0",
    "python-dotenv>=1.0.0",
    "ddgs>=6.0.0",
    "youtube-transcript-api>=0.6.0",
    "pypdf>=4.0.0",
    "feedparser>=6.0.12",
]

[project.optional-dependencies]
slack   = ["slack-bolt>=1.21.0"]
youtube = ["youtube-transcript-api>=0.6.0"]
search  = ["ddgs>=6.0.0"]
browse  = ["beautifulsoup4>=4.12.0"]
file    = ["pypdf>=4.0.0", "python-docx>=1.1.0"]
korean  = ["pyhwp>=0.1b15", "python-hwpx>=2.0"]
discord = ["discord.py>=2.0.0"]
rss     = ["feedparser>=6.0.0"]
all     = [
    "koclaw[slack]",
    "koclaw[youtube]",
    "koclaw[search]",
    "koclaw[browse]",
    "koclaw[file]",
    "koclaw[korean]",
    "koclaw[discord]",
    "koclaw[rss]",
]
[project.entry-points."koclaw.tools"]
# 설치된 tool이 자동으로 등록됨
# 예: youtube = "koclaw.tools.youtube:YouTubeTool"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.hatch.build.targets.wheel]
packages = ["koclaw"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.9.0",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I"]  # pycodestyle errors, pyflakes, isort
ignore = ["E501"]          # 줄 길이는 formatter에 위임
